added currentUserContext and connected editAccount api
